sched/fair: Prefer low capacity	idle-CPU for boosted non-prefer-idle tasks
The homescreen test-case shows unwanted disturbance on the big CPUs on
the Hikey620 platform with Android 4.9. There are multiple reasons for
that though.

By default boost and prefer_idle are enabled for both top-app and
foreground tasks and find_best_target() always (intentionally) prefers
the big CPUs if prefer_idle is enabled. And some of the foreground tasks
(like: DispSync, PowerManagerSer and PhotonicModulat) for the homescreen
test-case get placed on the big CPUs eventually because of that.

Even if prefer_idle is disabled for such foreground tasks, they don't
end up on the little CPUs. The reason being that find_best_target()
still prefers big CPUs if the task is boosted, though some of the
comments in find_best_target() routine say the exact opposite of that.
It eventually depends on the order in which CPUs are processed, which is
from big to little for boosted tasks.

This patch updates the find_best_target() routine to select low capacity
idle-CPU if the task is boosted but doesn't have prefer_idle enabled.
This would be the same for non-boosted tasks as well.

Signed-off-by: Leo Yan <leo.yan at linaro.org>
Signed-off-by: Viresh Kumar <viresh.kumar at linaro.org>
(cherry picked from commit 6b167aae670b375510894f4ac09bafe585256678)
(cherry picked from commit 7e3c1d13ba2ec6c7ec5512da098b00cf232ab3f9)
Signed-off-by: zachariasmaladroit <zacharias.maladroit@gmail.com>